+++ lib/admin/images.php
@@ -26,7 +26,7 @@ function mai_hide_featured_image_checkbox( $content, $post_id, $thumbnail_id ) {
 
 	// Build our new field
 	$new_field = sprintf( '<p><label for="mai_hide_featured_image"><input type="checkbox" id="mai_hide_featured_image" name="mai_hide_featured_image" %s>%s</label></p>',
-		checked( get_post_meta( $post_id, 'mai_hide_featured_image', true ), true, false ),
+		checked( (bool) mai_sanitize_one_zero( get_post_meta( $post_id, 'mai_hide_featured_image', true ) ), true, false ),
 		__( 'Hide featured image', 'mai-pro-engine' )
 	);
 
@@ -47,6 +47,10 @@ function mai_save_hide_featured_image_checkbox( $post_id, $post ) {
 
 	// Bail if we don't have a value at all
 	if ( ! isset( $_POST[ 'mai_hide_featured_image' ] ) ) {
+		$hide = get_post_meta( $post_id, 'mai_hide_featured_image', true );
+		if ( $hide ) {
+			delete_post_meta( $post_id, 'mai_hide_featured_image' );
+		}
 		return;
 	}
 
@@ -99,8 +103,8 @@ function mai_remove_unsupported_flexgrid_gallery_options() {
 		.gallery-settings .columns option[value="7"],
 		.gallery-settings .columns option[value="8"],
 		.gallery-settings .columns option[value="9"] {
-		    display:none !important;
-		    visibility: hidden !important;
+			display:none !important;
+			visibility: hidden !important;
 		}
 	</style>';
 }