+++ CHANGES.md
@@ -1,9 +1,15 @@
 ### Changelog
 
+#### 1.1.3 (9/14/17)
+* Fixed: Critical bug where saving Genesis > Theme Settings were resetting all custom settings.
+* Fixed: Hide Featured Image setting unable to save as unchecked after saving as checked.
+* Fixed: Hide Featured Image setting not hiding image on WooCommerce products.
+
 #### 1.1.2 (9/13/17)
 * Added: [grid] New "image_align" parameter. Accepts left, center, or right. This allows [grid] to display content exactly like default archives (e.g. the blog).
 * Added: [col] New "bg" parameter. Accepts hex value. Example: [col bg="#000000"].
 * Added: [grid] [col] New "bottom" parameter. This allows you to define the bottom margin (spacing) on each entry/column. Example: [col bottom="10"]. This would add 10px of margin to the bottom. Valid values are 0, 5, 10, 20, 30, 40, 50, 60.
+* Changed: Updated plugin-update-checker to latest version (4.2).
 * Fixed: Posts per page setting not working on CPT archives.
 * Fixed: Max width on entry pagination images when images weren't regenerated after activating Mai Pro.